+++ src/tar.c
@@ -248,7 +248,7 @@ static void add_pax_header(
   /* build the string */
   blob_appendf(&tball.pax, "%d %s=%*.*s\n", blen, zField, nValue, nValue, zValue);
   /* this _must_ be right */
-  if(blob_size(&tball.pax) != blen){
+  if((int)blob_size(&tball.pax) != blen){
     fossil_panic("internal error: PAX tar header has bad length");
   }
 }
